backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
twilio>=8.0.0
numpy>=1.26.0
pandas>=2.2.0
python-multipart>=0.0.9
appwrite>=9.0.0
edge-tts>=6.1.0
mistralai>=1.0.0
google-generativeai>=0.7.2
groq>=0.9.0
pymupdf>=1.24.0
